added getting price for each product from wb

+++ src/main/java/version_1/dto/WBResponseDtos/WBItemSize.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class WBItemSize {
     Integer sizeId;
     Integer price;
-    Float discountedPrice;
-    Float clubDiscountedPrice;
+    Double discountedPrice;
+    Double clubDiscountedPrice;
     String techSizeName;
 }
